+++ buession-springboot-datasource/src/main/java/com/buession/springboot/datasource/autoconfigure/DataSourceAutoConfiguration.java
@@ -19,7 +19,7 @@
  * +-------------------------------------------------------------------------------------------------------+
  * | License: http://www.apache.org/licenses/LICENSE-2.0.txt 										       |
  * | Author: Yong.Teng <webmaster@buession.com> 													       |
- * | Copyright @ 2013-2024 Buession.com Inc.														       |
+ * | Copyright @ 2013-2026 Buession.com Inc.														       |
  * +-------------------------------------------------------------------------------------------------------+
  */
 package com.buession.springboot.datasource.autoconfigure;
@@ -70,7 +70,7 @@ public class DataSourceAutoConfiguration {
 	@ConditionalOnMissingBean({DataSource.class, XADataSource.class})
 	@Import({DataSourceConfiguration.Dbcp2.class, DataSourceConfiguration.Druid.class,
 			DataSourceConfiguration.Hikari.class, DataSourceConfiguration.Oracle.class,
-			DataSourceConfiguration.Tomcat.class, DataSourceConfiguration.Generic.class})
+			DataSourceConfiguration.Tomcat.class})
 	static class PooledDataSourceConfiguration {
 
 	}