+++ buession-springboot-datasource/src/main/java/com/buession/springboot/datasource/autoconfigure/DataSourceConfiguration.java
@@ -31,9 +31,6 @@ import com.buession.jdbc.core.Callback;
 import com.buession.jdbc.datasource.*;
 import com.buession.jdbc.datasource.pool.*;
 import com.buession.springboot.datasource.core.DataSourceType;
-import com.buession.springboot.datasource.core.DynamicUrlBuilder;
-import oracle.ucp.jdbc.PoolDataSource;
-import org.apache.commons.dbcp2.BasicDataSource;
 import org.springframework.beans.factory.ObjectProvider;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
@@ -63,11 +60,11 @@ public class DataSourceConfiguration {
 
 	protected static <ODS extends javax.sql.DataSource, C extends BaseConfig, P extends PoolConfiguration, DS extends com.buession.jdbc.datasource.DataSource<ODS, P>> DataSource createDataSource(
 			final Class<DS> type, final DataSourceProperties dataSourceProperties, final C dataSourceConfig,
-			final P poolConfiguration, final DynamicUrlBuilder dynamicUrlBuilder, final Configurer<DS, C> customizer,
+			final P poolConfiguration, final Configurer<DS, C> customizer,
 			final Callback<ODS, DataSourceProperties> callback) {
 		final DataSourceInitializer<C, P, ODS, DS> dataSourceInitializer = new DataSourceInitializer<>(type,
 				dataSourceProperties, dataSourceConfig, poolConfiguration, customizer, callback);
-		return dataSourceInitializer.createDataSource(dynamicUrlBuilder);
+		return dataSourceInitializer.createDataSource();
 	}
 
 	/**
@@ -77,9 +74,9 @@ public class DataSourceConfiguration {
 	 */
 	@AutoConfiguration
 	@EnableConfigurationProperties(DataSourceProperties.class)
-	@ConditionalOnClass(BasicDataSource.class)
+	@ConditionalOnClass(name = DataSourceType.DHCP2)
 	@ConditionalOnMissingBean(DataSource.class)
-	@ConditionalOnProperty(prefix = DataSourceProperties.PREFIX, name = "type", havingValue = DataSourceType.DHCP2, matchIfMissing = true)
+	@ConditionalOnProperty(prefix = DataSourceProperties.PREFIX, name = "type", havingValue = DataSourceType.DHCP2)
 	static class Dbcp2 extends DataSourceConfiguration {
 
 		public Dbcp2(DataSourceProperties properties) {
@@ -88,9 +85,10 @@ public class DataSourceConfiguration {
 
 		@Bean
 		@ConfigurationProperties(prefix = DataSourceProperties.PREFIX + ".dbcp2")
-		public DataSource dataSource(ObjectProvider<DynamicUrlBuilder> dynamicUrlBuilder) {
+		public DataSource dataSource(ObjectProvider<DataSourceConsumer> dataSourceConsumer) {
+			dataSourceConsumer.orderedStream().forEach(consumer->consumer.accept(properties));
 			return createDataSource(Dbcp2DataSource.class, properties, properties.getDbcp2(),
-					new Dbcp2PoolConfiguration(), dynamicUrlBuilder.getIfAvailable(), (dataSource, config)->{
+					new Dbcp2PoolConfiguration(), (dataSource, config)->{
 						propertyMapper.from(config::getConnectionFactoryClassName)
 								.to(dataSource::setConnectionFactoryClassName);
 
@@ -108,7 +106,7 @@ public class DataSourceConfiguration {
 		}
 
 		private static void poolConfig(final Dbcp2PoolConfiguration poolConfiguration,
-		                               final Dbcp2Config dataSourceConfig) {
+									   final Dbcp2Config dataSourceConfig) {
 			propertyMapper.from(dataSourceConfig::getMaxConnLifetime).to(poolConfiguration::setMaxConnLifetime);
 
 			propertyMapper.from(dataSourceConfig::getPoolPreparedStatements)
@@ -143,9 +141,9 @@ public class DataSourceConfiguration {
 	 */
 	@AutoConfiguration
 	@EnableConfigurationProperties(DataSourceProperties.class)
-	@ConditionalOnClass(DruidDataSource.class)
+	@ConditionalOnClass(name = DataSourceType.DRUID)
 	@ConditionalOnMissingBean(DataSource.class)
-	@ConditionalOnProperty(prefix = DataSourceProperties.PREFIX, name = "type", havingValue = DataSourceType.DRUID, matchIfMissing = true)
+	@ConditionalOnProperty(prefix = DataSourceProperties.PREFIX, name = "type", havingValue = DataSourceType.DRUID)
 	static class Druid extends DataSourceConfiguration {
 
 		public Druid(DataSourceProperties properties) {
@@ -154,9 +152,10 @@ public class DataSourceConfiguration {
 
 		@Bean
 		@ConfigurationProperties(prefix = DataSourceProperties.PREFIX + ".druid")
-		public DataSource dataSource(ObjectProvider<DynamicUrlBuilder> dynamicUrlBuilder) {
+		public DataSource dataSource(ObjectProvider<DataSourceConsumer> dataSourceConsumer) {
+			dataSourceConsumer.orderedStream().forEach(consumer->consumer.accept(properties));
 			return createDataSource(DruidDataSource.class, properties, properties.getDruid(),
-					new DruidPoolConfiguration(), dynamicUrlBuilder.getIfAvailable(), (dataSource, config)->{
+					new DruidPoolConfiguration(), (dataSource, config)->{
 						propertyMapper.from(config::getUserCallbackClassName).to(dataSource::setUserCallbackClassName);
 						propertyMapper.from(config::getPasswordCallbackClassName)
 								.to(dataSource::setPasswordCallbackClassName);
@@ -219,7 +218,7 @@ public class DataSourceConfiguration {
 		}
 
 		private static void poolConfig(final DruidPoolConfiguration poolConfiguration,
-		                               final DruidConfig dataSourceConfig) {
+									   final DruidConfig dataSourceConfig) {
 			propertyMapper.from(dataSourceConfig::getMaxActive).to(poolConfiguration::setMaxActive);
 
 			propertyMapper.from(dataSourceConfig::getKeepAlive).to(poolConfiguration::setKeepAlive);
@@ -269,9 +268,9 @@ public class DataSourceConfiguration {
 	 */
 	@AutoConfiguration
 	@EnableConfigurationProperties(DataSourceProperties.class)
-	@ConditionalOnClass(HikariDataSource.class)
+	@ConditionalOnClass(name = DataSourceType.HIKARI)
 	@ConditionalOnMissingBean(DataSource.class)
-	@ConditionalOnProperty(prefix = DataSourceProperties.PREFIX, name = "type", havingValue = DataSourceType.HIKARI, matchIfMissing = true)
+	@ConditionalOnProperty(prefix = DataSourceProperties.PREFIX, name = "type", havingValue = DataSourceType.HIKARI)
 	static class Hikari extends DataSourceConfiguration {
 
 		public Hikari(DataSourceProperties properties) {
@@ -280,9 +279,10 @@ public class DataSourceConfiguration {
 
 		@Bean
 		@ConfigurationProperties(prefix = DataSourceProperties.PREFIX + ".hikari")
-		public DataSource dataSource(ObjectProvider<DynamicUrlBuilder> dynamicUrlBuilder) {
+		public DataSource dataSource(ObjectProvider<DataSourceConsumer> dataSourceConsumer) {
+			dataSourceConsumer.orderedStream().forEach(consumer->consumer.accept(properties));
 			return createDataSource(HikariDataSource.class, properties, properties.getHikari(),
-					new HikariPoolConfiguration(), dynamicUrlBuilder.getIfAvailable(), (dataSource, config)->{
+					new HikariPoolConfiguration(), (dataSource, config)->{
 						propertyMapper.from(config::getJndiName).to(dataSource::setJndiName);
 
 						propertyMapper.from(config::getConnectionTimeout).to(dataSource::setConnectionTimeout);
@@ -295,7 +295,7 @@ public class DataSourceConfiguration {
 		}
 
 		private static void poolConfig(final HikariPoolConfiguration poolConfiguration,
-		                               final HikariConfig dataSourceConfig) {
+									   final HikariConfig dataSourceConfig) {
 			propertyMapper.from(dataSourceConfig::getInitializationFailTimeout)
 					.to(poolConfiguration::setInitializationFailTimeout);
 
@@ -332,9 +332,9 @@ public class DataSourceConfiguration {
 	 */
 	@AutoConfiguration
 	@EnableConfigurationProperties(DataSourceProperties.class)
-	@ConditionalOnClass(PoolDataSource.class)
+	@ConditionalOnClass(name = DataSourceType.ORACLE)
 	@ConditionalOnMissingBean(DataSource.class)
-	@ConditionalOnProperty(prefix = DataSourceProperties.PREFIX, name = "type", havingValue = DataSourceType.ORACLE, matchIfMissing = true)
+	@ConditionalOnProperty(prefix = DataSourceProperties.PREFIX, name = "type", havingValue = DataSourceType.ORACLE)
 	static class Oracle extends DataSourceConfiguration {
 
 		public Oracle(DataSourceProperties properties) {
@@ -343,9 +343,10 @@ public class DataSourceConfiguration {
 
 		@Bean
 		@ConfigurationProperties(prefix = DataSourceProperties.PREFIX + ".oracle")
-		public DataSource dataSource(ObjectProvider<DynamicUrlBuilder> dynamicUrlBuilder) {
+		public DataSource dataSource(ObjectProvider<DataSourceConsumer> dataSourceConsumer) {
+			dataSourceConsumer.orderedStream().forEach(consumer->consumer.accept(properties));
 			return createDataSource(OracleDataSource.class, properties, properties.getOracle(),
-					new OraclePoolConfiguration(), dynamicUrlBuilder.getIfAvailable(), (dataSource, config)->{
+					new OraclePoolConfiguration(), (dataSource, config)->{
 						propertyMapper.from(config::getNetworkProtocol).to(dataSource::setNetworkProtocol);
 						propertyMapper.from(config::getServerName).to(dataSource::setServerName);
 						propertyMapper.from(config::getPortNumber).to(dataSource::setPortNumber);
@@ -376,7 +377,7 @@ public class DataSourceConfiguration {
 		}
 
 		private static void poolConfig(final OraclePoolConfiguration poolConfiguration,
-		                               final OracleConfig dataSourceConfig) {
+									   final OracleConfig dataSourceConfig) {
 			propertyMapper.from(dataSourceConfig::getMinPoolSize).to(poolConfiguration::setMinPoolSize);
 			propertyMapper.from(dataSourceConfig::getMaxPoolSize).to(poolConfiguration::setMaxPoolSize);
 
@@ -422,9 +423,9 @@ public class DataSourceConfiguration {
 	 */
 	@AutoConfiguration
 	@EnableConfigurationProperties(DataSourceProperties.class)
-	@ConditionalOnClass(org.apache.tomcat.jdbc.pool.DataSource.class)
+	@ConditionalOnClass(name = DataSourceType.TOMCAT)
 	@ConditionalOnMissingBean(DataSource.class)
-	@ConditionalOnProperty(prefix = DataSourceProperties.PREFIX, name = "type", havingValue = DataSourceType.TOMCAT, matchIfMissing = true)
+	@ConditionalOnProperty(prefix = DataSourceProperties.PREFIX, name = "type", havingValue = DataSourceType.TOMCAT)
 	static class Tomcat extends DataSourceConfiguration {
 
 		public Tomcat(DataSourceProperties properties) {
@@ -433,9 +434,10 @@ public class DataSourceConfiguration {
 
 		@Bean
 		@ConfigurationProperties(prefix = DataSourceProperties.PREFIX + ".tomcat")
-		public DataSource dataSource(ObjectProvider<DynamicUrlBuilder> dynamicUrlBuilder) {
+		public DataSource dataSource(ObjectProvider<DataSourceConsumer> dataSourceConsumer) {
+			dataSourceConsumer.orderedStream().forEach(consumer->consumer.accept(properties));
 			return createDataSource(TomcatDataSource.class, properties, properties.getTomcat(),
-					new TomcatPoolConfiguration(), dynamicUrlBuilder.getIfAvailable(), (dataSource, config)->{
+					new TomcatPoolConfiguration(), (dataSource, config)->{
 						propertyMapper.from(config::getJndiName).to(dataSource::setJndiName);
 
 						propertyMapper.from(config::getAlternateUsernameAllowed)
@@ -464,7 +466,7 @@ public class DataSourceConfiguration {
 		}
 
 		private static void poolConfig(final TomcatPoolConfiguration poolConfiguration,
-		                               final TomcatConfig dataSourceConfig) {
+									   final TomcatConfig dataSourceConfig) {
 			propertyMapper.from(dataSourceConfig::getMaxActive).to(poolConfiguration::setMaxActive);
 			propertyMapper.from(dataSourceConfig::getMaxAge).to(poolConfiguration::setMaxAge);
 
@@ -489,31 +491,6 @@ public class DataSourceConfiguration {
 
 			propertyMapper.from(dataSourceConfig::getUseLock).to(poolConfiguration::setUseLock);
 			propertyMapper.from(dataSourceConfig::getUseEquals).to(poolConfiguration::setUseEquals);
-		}
-
-	}
-
-	/**
-	 * Generic DataSource Configuration.
-	 *
-	 * @since 1.3.2
-	 */
-	@AutoConfiguration
-	@EnableConfigurationProperties(DataSourceProperties.class)
-	@ConditionalOnMissingBean(DataSource.class)
-	@ConditionalOnProperty(prefix = DataSourceProperties.PREFIX, name = "type")
-	static class Generic extends DataSourceConfiguration {
-
-		public Generic(DataSourceProperties properties) {
-			super(properties);
-		}
-
-		@Bean
-		public DataSource dataSource(ObjectProvider<DynamicUrlBuilder> dynamicUrlBuilder) {
-			return createDataSource(GenericDataSource.class, properties, properties.getGeneric(), null,
-					dynamicUrlBuilder.getIfAvailable(), (dataSource, config)->{
-
-					}, (dataSource, config)->dataSource);
 		}
 
 	}